refactor(file-service): share one upload engine for versioned resources
The multipart upload machinery in DatasetResource — session init, part
upload, finish, abort, list, the one-shot upload and staged-file delete —
is resource-agnostic apart from the columns it names. Model upload needs the
same machinery, so move it behind a descriptor instead of copying ~900 lines.

ResourceStorage names a resource's LakeFS repository column plus its
*_upload_session and *_upload_session_part tables; ResourceUploadService
holds the single implementation. DatasetResource keeps its endpoints and
DTOs and delegates the logic through ResourceStorage.Datasets.

Behavior is unchanged: the engine is the dataset implementation moved
verbatim with its column references parameterized, and transactions keep
their existing boundaries. The locking protocol (FOR UPDATE NOWAIT, SQLState
55P03 to 409), the part-size arithmetic and its overflow guards, ETag
idempotency and the resume/restart rules are now defined once.

Part of #6494, groundwork for #6498.
